backend/requirements.txt
flask
flask-cors
pymongo
pandas
numpy
scikit-learn
joblib
pyjwt
passlib[bcrypt]